Split builder into separate files, fixed up imports, added copyrights
